feat: add sticky banner to community page
