+++ index.html
@@ -78,7 +78,12 @@
   </head>
 
   <body class="bg-[#1A1A1D] text-white font-sans flex flex-col">
-    <div id="theme-banner" class="bg-[#30D5C8] text-black text-center py-1 hidden">24% off when you order 3 prints</div>
+    <div
+      id="theme-banner"
+      class="bg-[#30D5C8] text-black text-center py-1 hidden sticky top-0 z-50"
+    >
+      24% off when you order 3 prints
+    </div>
 
     <!-- HEADER (unchanged UI) -->
     <header class="relative flex items-center justify-between py-4 px-6">
